% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_Andes.R
\docType{data}
\name{Andes}
\alias{Andes}
\title{Occurrence data for the class Amphibia in the Andes region from 2011 to 2022}
\format{
A data frame with 118267 rows and 195 columns:
}
\source{
Queried from the Global Biodiversity Information Facility database.
}
\usage{
Andes
}
\description{
Contains taxa and location data and many other occurrence variables.
}
\examples{
data("Andes")
}
\keyword{datasets}
