% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_Colombia.R
\docType{data}
\name{Colombia}
\alias{Colombia}
\title{Occurrence data of the Formicidae family in Colombia from 2000 to 2023}
\format{
A data frame with 190810 rows and 188 columns:
}
\source{
Queried from the Global Biodiversity Information Facility database.
}
\usage{
Colombia
}
\description{
Contains taxa and location data and many other occurence variables.
}
\examples{
data("Colombia")
}
\keyword{datasets}
