% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DivInsight-package.R
\docType{package}
\name{DivInsight-package}
\alias{DivInsight}
\alias{DivInsight-package}
\title{DivInsight: Diversity Index Calculation & Visualisation for Taxa and Location}
\description{
The functions in the DivInsight package can be used to repurpose occurrence data for calculating diversity index values, creating visuals, and generating species composition matrices for a chosen taxon and location.
}
\keyword{internal}
