% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_groups_by_minsize.R
\name{filter_groups_by_minsize}
\alias{filter_groups_by_minsize}
\title{Filter Group by Size}
\usage{
filter_groups_by_minsize(clusterised_object, min_group_size)
}
\arguments{
\item{clusterised_object}{An object created by the clusterise_sites functions.}

\item{min_group_size}{The minimum size of each group.}
}
\value{
A new clusterised object with groups of at least the specified minimum size.
}
\description{
Filter a clusterised object by removing groups below a specified size.
}
\examples{
# return groups that have at least 10 sites
\dontrun{filter_groups_by_minsize(

clusterised_object = clusterised_Meta_20km_groups,
min_group_size = 10

)}
}
