% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_groups_by_number.R
\name{filter_groups_by_number}
\alias{filter_groups_by_number}
\title{Filter Group by Name}
\usage{
filter_groups_by_number(clusterised_object, group_number)
}
\arguments{
\item{clusterised_object}{An object created by the clusterise_sites function.}

\item{group_number}{A number that identifies the group to be extracted.}
}
\value{
A new clusterised object that only contains sites from the specified group.
}
\description{
Filter a clusterised object that has been assigned groups and remove groups that are not a specified label.
}
\examples{
# extract groups 6, 5, and 1 from a clusterised object
\dontrun{Colombia_Meta_6 <- filter_groups_by_name(clusterised_Meta_20km_groups, 6)}
\dontrun{Colombia_Meta_5 <- filter_groups_by_name(clusterised_Meta_20km_groups, 5)}
\dontrun{Colombia_Meta_1 <- filter_groups_by_name(clusterised_Meta_20km_groups, 1)}
}
