% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_taxon_range.R
\name{find_taxon_range}
\alias{find_taxon_range}
\title{Find Taxon Range}
\usage{
find_taxon_range(taxon_dataframe)
}
\arguments{
\item{taxon_dataframe}{A dataset containing occurrence data for the taxon of interest.}
}
\value{
Returns a numeric value representing the maximum distance (metres) between all individuals within the given dataset.
}
\description{
Calculate the distances between all individuals within a dataset and return the maximum linear distance.
}
\examples{
# find maximum taxon range in different years for the Centrolenidae family
Centro_2011 <- subset(Andes, family == "Centrolenidae" & year == 2011)
Centro_2016 <- subset(Andes, family == "Centrolenidae" & year == 2016)
Centro_2021 <- subset(Andes, family == "Centrolenidae" & year == 2021)

find_taxon_range(Centro_2011)
find_taxon_range(Centro_2016)
find_taxon_range(Centro_2021)
}
