% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbif_to_dataframe.R
\name{gbif_to_dataframe}
\alias{gbif_to_dataframe}
\title{Convert GBIF Object}
\usage{
gbif_to_dataframe(GBIF_object)
}
\arguments{
\item{GBIF_object}{An object queried from the GBIF database: a list containing occurrence data for a specified taxon and location.}
}
\value{
An R dataframe containing the occurrence data from the GBIF object.
}
\description{
Convert a GBIF object into workable R dataframe.
}
\examples{
# convert the GBIF object containing occurrence data from 2000 to 2004
\dontrun{df_data_Ants_Colombia_2000_to_2004 <- gbif_to_dataframe(
 GBIF_data_Ants_Colombia_2000_to_2004
)}
# convert the GBIF object containing occurence data from 2005 to 2008
\dontrun{df_data_Ants_Colombia_2005_to_2008 <- gbif_to_dataframe(
 GBIF_data_Ants_Colombia_2005_to_2008
)}
}
