% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_spec_tables.R
\name{generate_spec_tables}
\alias{generate_spec_tables}
\title{Generate Species Tables}
\usage{
generate_spec_tables(clusterised_object, min_individuals = 0, min_species = 0)
}
\arguments{
\item{clusterised_object}{An object created by the clusterise_sites function.}

\item{min_individuals}{The minimum number of individuals in each table}

\item{min_species}{The minimum number of species in each table.}
}
\value{
A list of species counts tables.
}
\description{
Generate a list of tables for species counts from the input clusterised object.
}
\examples{
# generate species count tables using a clusterised object.
\dontrun{SCM_Santander_1 <- generate_speccomm(

 clusterised_object = Colombia_Santander_1,

 min_individuals = 30,
 min_species = 5

)}
}
