% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_spec_tables_df.R
\name{generate_spec_tables_df}
\alias{generate_spec_tables_df}
\title{Generate Species Tables}
\usage{
generate_spec_tables_df(dataframe, min_individuals = 0, min_species = 0)
}
\arguments{
\item{dataframe}{An occurrence dataframe.}

\item{min_individuals}{The minimum number of individuals in each table}

\item{min_species}{The minimum number of species in each table.}
}
\value{
A list of species counts tables.
}
\description{
Generate a list of tables for species counts from the input dataframe.
}
\examples{
# generate species count tables for the entire country of Colombia.
\dontrun{SCM_Colombia <- generate_speccomm_df(

 dataframe = Colombia,

 min_individuals = 50,
 min_species = 10

)}
}
