% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_speccomm.R
\name{generate_speccomm}
\alias{generate_speccomm}
\title{Generate Species Composition Matrix}
\usage{
generate_speccomm(spec_table_list)
}
\arguments{
\item{spec_table_list}{A list of species tables created by the generate_spec_tables function.}
}
\value{
A species composition matrix
}
\description{
Create a species composition matrix using a list of species tables.
}
\examples{
# create a SCM using a list of species
\dontrun{SCM <- generate_speccomm(comb.df)}
}
