% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_stats.R
\name{generate_stats}
\alias{generate_stats}
\title{Generate Diversity Index Statistics}
\usage{
generate_stats(clusterised_object)
}
\arguments{
\item{clusterised_object}{An object created by the clusterise_sites function.}
}
\value{
A dataframe where each row pertains to a site and each column has time & location data as well as diversity index values.
}
\description{
Calculates diversity index values for each site in a clusterised object.
}
\examples{
# generate statistics using clusterised data from the Meta province of Colombia
\dontrun{stats_Meta <- generate_stats(clusterised_Meta)}
}
