% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_add.R
\name{map_add}
\alias{map_add}
\title{Add to Site Map}
\usage{
map_add(existing_map, clusterised_object, site_name = "", colour = "#09f")
}
\arguments{
\item{existing_map}{A map object that has been created using the map_start function.}

\item{clusterised_object}{An object created by the clusterise_sites functions.}

\item{site_name}{An optional label to add the the coordinate pins.}

\item{colour}{The colour of the coordinate pins.}
}
\value{
Coordinate pins to be added to an existing interactive map object.
}
\description{
Add coordinate pins to an existing interactive map.
}
\examples{
# add coordinate pins to an existing map object
\dontrun{Colombia_Meta_map <-

 map_add(
   existing_map = Colombia_Meta_map,

   clusterised_object = Colombia_Meta_5,
   site_name = "Meta#5",
   colour = "purple"
 )}
}
