% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_start.R
\name{map_start}
\alias{map_start}
\title{Create Site Map}
\usage{
map_start(clusterised_object, site_name = "", colour = "#60f")
}
\arguments{
\item{clusterised_object}{An object created by the clusterise_sites function.}

\item{site_name}{An optional label to add to the coordinate pins.}

\item{colour}{The colour of the coordinate pins.}
}
\value{
A Viewer object: an interactive map.
}
\description{
Create interactive map using centred site coordinates.
}
\examples{
# create a new map using coordinates from site 1 in the Meta province of Colombia
\dontrun{Colombia_Meta_map <-

 map_start(
   clusterised_object = Colombia_Meta_1,
   site_name = "Meta#1",
   colour = "green"
 )}
}
