% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapAdd_indivCirc.R
\name{mapAdd_indivCirc}
\alias{mapAdd_indivCirc}
\title{Add to Individual Occurrence Map}
\usage{
mapAdd_indivCirc(
  existing_map,
  taxon_dataframe,
  site_name = "",
  colour = "#60f"
)
}
\arguments{
\item{existing_map}{A map object that has been created using the map_start function.}

\item{site_name}{An optional label to add the the coordinate pins.}

\item{colour}{The colour of the coordinate pins.}

\item{clusterised_object}{An object created by the clusterise_sites functions.}
}
\value{
Coordinate pins to be added to an existing interactive map object.
}
\description{
Add coordinate circles to an existing interactive map.
}
\examples{
# add site coordinates to an existing map
dataframe_Lepto <- subset(Andes, family == "Leptodactylidae")

map <- mapStart_indivCirc(dataframe_Lepto, site_name = "Whole Dataset")

dataframe_Lepto_split <- split(dataframe_Lepto, dataframe_Lepto$year)

map <- mapAdd_indivCirc(map, dataframe_Lepto_split$`2011`, "2011", "red")
map <- mapAdd_indivCirc(map, dataframe_Lepto_split$`2016`, "2016", "purple")
map <- mapAdd_indivCirc(map, dataframe_Lepto_split$`2021`, "2021", "green")

map
}
