% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapAdd_rangePoly.R
\name{mapAdd_rangePoly}
\alias{mapAdd_rangePoly}
\title{Add to Individual Occurrence Map}
\usage{
mapAdd_rangePoly(
  existing_map,
  taxon_dataframe,
  site_name = "",
  colour = "#60f",
  site_legend = T,
  poly_opacity = 0.4
)
}
\arguments{
\item{existing_map}{A map object that has been created using the map_start function.}

\item{site_name}{An optional label to add the the coordinate pins.}

\item{colour}{The colour of the coordinate pins.}

\item{site_legend}{A boolean value to choose to add a site name legend to the map.}

\item{poly_opacity}{An optional value to set the opacity of the polygon.}

\item{clusterised_object}{An object created by the clusterise_sites functions.}
}
\value{
Coordinate pins to be added to an existing interactive map object.
}
\description{
Add range polygons to an existing interactive map.
}
\examples{
# add site coordinates to an existing map
species_dataframe <- subset(Andes, species == "Boana crepitans")

map <- mapStart_rangePoly(species_dataframe, site_name = "Whole Dataset")

species_dataframe_split <- split(species_dataframe, species_dataframe$year)

map <- mapAdd_rangePoly(map, species_dataframe_split$`2011`, "2016", "purple")
map <- mapAdd_rangePoly(map, species_dataframe_split$`2013`, "2013", "green")
map <- mapAdd_rangePoly(map, species_dataframe_split$`2015`, "2015", "red")
map <- mapAdd_rangePoly(map, species_dataframe_split$`2017`, "2017", "blue")

map
}
