% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapStart_rangePoly.R
\name{mapStart_rangePoly}
\alias{mapStart_rangePoly}
\title{Create Taxon Range Map}
\usage{
mapStart_rangePoly(
  taxon_dataframe,
  site_name = "",
  colour = "black",
  site_legend = T,
  poly_opacity = 0.2
)
}
\arguments{
\item{taxon_dataframe}{A dataframe with occurrence data for the chosen taxon.}

\item{site_name}{An optional label to add to the coordinate circles.}

\item{colour}{The colour of the coordinate circles.}

\item{site_legend}{A boolean value to choose to add a site name legend to the map.}

\item{poly_opacity}{An optional value to set the opacity of the polygon.}
}
\value{
A Viewer object: an interactive map.
}
\description{
Create interactive map and plot a taxon range polygon.
}
\examples{
# create an interactive map using coordinates from a taxon dataframe
species_dataframe <- subset(Andes, species == "Boana crepitans")

map <- mapStart_rangePoly(species_dataframe, site_name = "Whole Dataset")

map
}
