% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapStart_siteCirc.R
\name{mapStart_siteCirc}
\alias{mapStart_siteCirc}
\title{Create Site Map}
\usage{
mapStart_siteCirc(clusterised_object, site_name = "", colour = "#60f")
}
\arguments{
\item{clusterised_object}{An object created by the clusterise_sites function.}

\item{site_name}{An optional label to add to the coordinate circles.}

\item{colour}{The colour of the coordinate circles.}
}
\value{
A Viewer object: an interactive map.
}
\description{
Create interactive map using centred site coordinates.
}
\examples{
# create an interactive map using one of the group's coordinates
Colombia_Caquetá_dataframe <- subset(Colombia, stateProvince == "Caquetá")

clusterised_Caquetá <- clusterise_sites(dataframe = Colombia_Caquetá_dataframe,
                                       cluster_min_length = 30,
                                       group_radius = 20000
)

print(clusterised_Caquetá[[2]])

Colombia_Caquetá_3 <- filter_groups_by_number(clusterised_Caquetá, 3)

map <-

 mapStart_siteCirc(

   clusterised_object = Colombia_Caquetá_3,
   site_name = "Caquetá#3",
   colour = "green"

 )

 map
}
