% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sites_scatter_H.R
\name{plot_sites_scatter_H}
\alias{plot_sites_scatter_H}
\title{Plot H against Date}
\usage{
plot_sites_scatter_H(
  clusterised_object,
  xlabel = "Date",
  ylabel = "Diversity (H)",
  main_title = "Shannon's H over Time"
)
}
\arguments{
\item{clusterised_object}{An object created by the clusterise_sites function.}

\item{xlabel}{The X axis label with "Date" as default.}

\item{ylabel}{The Y axis label with "Diversity (H)" as default.}

\item{main_title}{The main label with "Shannon's H over Time" as default.}
}
\description{
Plot Shannon's H on the Y axis and dates on the X axis with scaled spacing for each date and scattered points.
}
\examples{
# plot scatter charts
\dontrun{plot_sites_scatter_H(Colombia_Meta_6, main = "Shannons H at Meta#6")}
}
