% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sites_scatter_N.R
\name{plot_sites_scatter_N}
\alias{plot_sites_scatter_N}
\title{Plot Local Abundance against Date}
\usage{
plot_sites_scatter_N(
  clusterised_object,
  xlabel = "Date",
  ylabel = "Abundance (N)",
  main_title = "Local Abundance over Time"
)
}
\arguments{
\item{clusterised_object}{An object created by the clusterise_sites function.}

\item{xlabel}{The X axis label with "Date" as default.}

\item{ylabel}{The Y axis label with "Abundance (N)" as default.}

\item{main_title}{The main label with "Local Abundance Over Time" as default.}
}
\value{
Returns a ggplot2 scatter plot in the plot panel, showing the relationship between dates and local abundance.
}
\description{
Plot local abundance on the Y axis and dates on the X axis with scaled spacing for each date and scattered points.
}
\examples{
# create a scatter plot showing abundance values over time at the Caquetá province
Colombia_Caquetá_dataframe <- subset(Colombia, stateProvince == "Caquetá")

clusterised_Caquetá <- clusterise_sites(dataframe = Colombia_Caquetá_dataframe,
                                       cluster_min_length = 30,
                                       group_radius = 20000
)

plot_sites_scatter_N(clusterised_Caquetá)
}
