% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sites_scatter_S.R
\name{plot_sites_scatter_S}
\alias{plot_sites_scatter_S}
\title{Plot Species Richness against Date}
\usage{
plot_sites_scatter_S(
  clusterised_object,
  xlabel = "Date",
  ylabel = "Species Richness (S)",
  main_title = "Species Richness over Time"
)
}
\arguments{
\item{clusterised_object}{An object created by the clusterise_sites function.}

\item{xlabel}{The X axis label with "Date" as default.}

\item{ylabel}{The Y axis label with "Species Richness (S)" as default.}

\item{main_title}{The main label with "Species Richness Over Time" as default.}
}
\value{
Returns a ggplot2 scatter plot in the plot panel, showing the relationship between dates and diversity index values.
}
\description{
Plot species richness on the Y axis and dates on the X axis with scaled spacing for each date and scattered points.
}
\examples{
# create a scatter plot showing species richness values over time at the Caquetá province
Colombia_Caquetá_dataframe <- subset(Colombia, stateProvince == "Caquetá")

clusterised_Caquetá <- clusterise_sites(dataframe = Colombia_Caquetá_dataframe,
                                       cluster_min_length = 30,
                                       group_radius = 20000
)

plot_sites_scatter_S(clusterised_Caquetá)
}
