% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sites_trend_N.R
\name{plot_sites_trend_N}
\alias{plot_sites_trend_N}
\title{Plot Local Abundance against Date}
\usage{
plot_sites_trend_N(
  clusterised_object,
  xlabel = "",
  ylabel = "Abundance (N)",
  main_title = "Local Abundance over Time"
)
}
\arguments{
\item{clusterised_object}{An object created by the clusterise_sites function.}

\item{xlabel}{The X axis label.}

\item{ylabel}{The Y axis label with "Abundance (N)" as default.}

\item{main_title}{The main title with "Local Abundance Over Time" as default.}
}
\value{
Returns a base R plot in the plot panel, depicting the trend of local abundance over time.
}
\description{
Plot local abundance on the Y axis and dates on the X axis with uniform spacing for each date and a line of best fit.
}
\examples{
# create a trend chart showing abundance values over time at the Caquetá province
Colombia_Caquetá_dataframe <- subset(Colombia, stateProvince == "Caquetá")

clusterised_Caquetá <- clusterise_sites(dataframe = Colombia_Caquetá_dataframe,
                                       cluster_min_length = 30,
                                       group_radius = 20000
)

plot_sites_trend_N(clusterised_Caquetá)
}
