% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sites_trend_S.R
\name{plot_sites_trend_S}
\alias{plot_sites_trend_S}
\title{Plot Species Richness against Date}
\usage{
plot_sites_trend_S(
  clusterised_object,
  xlabel = "",
  ylabel = "Diversity (S)",
  main_title = "Species Richness over Time"
)
}
\arguments{
\item{clusterised_object}{An object created by the clusterise_sites function.}

\item{xlabel}{The X axis label.}

\item{ylabel}{The Y axis label with "Species Richness (S)" as default.}

\item{main_title}{The main title with "Species Richness Over Time" as default.}
}
\value{
Returns a base R plot in the plot panel, depicting the trend of diversity index values over time.
}
\description{
Plot Species Richness on the Y axis and dates on the X axis with uniform spacing for each date and a line of best fit.
}
\examples{
# create a trend chart showing species richness values over time at the Caquetá province
Colombia_Caquetá_dataframe <- subset(Colombia, stateProvince == "Caquetá")

clusterised_Caquetá <- clusterise_sites(dataframe = Colombia_Caquetá_dataframe,
                                       cluster_min_length = 30,
                                       group_radius = 20000
)

plot_sites_trend_S(clusterised_Caquetá)
}
