% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_taxon_range.R
\name{plot_taxon_range}
\alias{plot_taxon_range}
\title{Plot Taxon Range against Year}
\usage{
plot_taxon_range(taxon_dataframe, main_title = "", abundance_col = "black")
}
\arguments{
\item{taxon_dataframe}{A dataset containing occurrence data for the taxon of interest.}

\item{abundance_col}{A character value that determines the colour of the abundance annotations.}

\item{title}{A character value to change the main title of the char}
}
\value{
Returns a ggplot2 bar plot, showing the relationship between years and taxon range.
}
\description{
Display the the maximum inter-individual distance and abundance of a specified taxon over time.
}
\examples{
# plot abundance and maximum taxon range for the Centrolenidae family
Centro <- subset(Andes, family == "Centrolenidae")

plot_taxon_range(Centro, "darkred")
}
