% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_climate_data.R
\name{query_climate_data}
\alias{query_climate_data}
\title{Query Climate Data}
\usage{
query_climate_data(stats_dataframe, api_key)
}
\arguments{
\item{stats_dataframe}{A dataframe created by using the `generate_stats()` function.}

\item{api_key}{An API key acquired by signing up to `stormglass.io`.}
}
\value{
A nested list containing elevation data and query metadata for the given stats and coordinates.
}
\description{
Query climate data from the Storm Glass datagrid.
}
\examples{
# query elevation data using Singapore data clusters
\dontrun{

stats_df <- generate_stats(Singpore_clst)

API <- "a1b2c3d4e5"

data_elev <- query_elevation_data(stats_df, API)

}
}
