% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Colombia}
\alias{Colombia}
\title{Occurence data of the Formicidae family in Colombia from 2000 to 2023}
\format{
## `Colombia`
A data frame with 190810 rows and 188 columns:
\describe{
  \item{decimalLongitude}{Longitude of the occurence}
  \item{decimalLatitude}{Latitude of the occurence}
  ...
}
}
\source{
Queried from the Global Biodiversity Information Facility database.
}
\usage{
Colombia
}
\description{
Contains taxa and location data and many other occurence variables
}
\keyword{datasets}
